+++ mirrord-agent/src/main.rs
@@ -3,7 +3,7 @@
 #![feature(once_cell)]
 
 use std::{
-    collections::{HashMap, HashSet},
+    collections::HashSet,
     net::{Ipv4Addr, SocketAddrV4},
     path::PathBuf,
 };
@@ -19,12 +19,11 @@ use futures::{
 };
 use mirrord_protocol::{
     tcp::{DaemonTcp, LayerTcp, LayerTcpSteal},
-    ClientMessage, DaemonCodec, DaemonMessage, GetEnvVarsRequest, RemoteResult,
+    ClientMessage, DaemonCodec, DaemonMessage, GetEnvVarsRequest,
 };
 use outgoing::{udp::UdpOutgoingApi, TcpOutgoingApi};
 use sniffer::{SnifferCommand, TCPConnectionSniffer, TCPSnifferAPI};
 use tokio::{
-    io::AsyncReadExt,
     net::{TcpListener, TcpStream},
     select,
     sync::mpsc::{self, Receiver, Sender},
@@ -41,6 +40,7 @@ use crate::{
 
 mod cli;
 mod dns;
+mod env;
 mod error;
 mod file;
 mod outgoing;
@@ -73,63 +73,6 @@ impl State {
         self.clients.remove(&client_id);
         self.index_allocator.free_index(client_id)
     }
-}
-
-/// Helper function that loads the process' environment variables, and selects only those that were
-/// requested from `mirrord-layer` (ignores vars specified in `filter_env_vars`).
-///
-/// Returns an error if none of the requested environment variables were found.
-async fn select_env_vars(
-    environ_path: PathBuf,
-    filter_env_vars: HashSet<String>,
-    select_env_vars: HashSet<String>,
-) -> RemoteResult<HashMap<String, String>> {
-    trace!(
-        "select_env_vars -> environ_path {:#?} filter_env_vars {:#?} select_env_vars {:#?}",
-        environ_path,
-        filter_env_vars,
-        select_env_vars
-    );
-
-    let mut environ_file = tokio::fs::File::open(environ_path).await?;
-
-    let mut raw_env_vars = String::with_capacity(8192);
-
-    // TODO: nginx doesn't play nice when we do this, it only returns a string that goes like
-    // "nginx -g daemon off;".
-    let _read_amount = environ_file.read_to_string(&mut raw_env_vars).await?;
-
-    // TODO: These are env vars that should usually be ignored. Revisit this list if a user
-    // ever asks for a way to NOT filter out these.
-    let mut default_filter = HashSet::with_capacity(2);
-    default_filter.insert("PATH".to_string());
-    default_filter.insert("HOME".to_string());
-
-    let env_vars = raw_env_vars
-        // "DB=foo.db\0PORT=99\0HOST=\0PATH=/fake\0"
-        .split_terminator(char::from(0))
-        // ["DB=foo.db", "PORT=99", "HOST=", "PATH=/fake"]
-        .map(|key_and_value| key_and_value.splitn(2, '=').collect::<Vec<_>>())
-        // [["DB", "foo.db"], ["PORT", "99"], ["HOST"], ["PATH", "/fake"]]
-        .filter_map(
-            |mut keys_and_values| match (keys_and_values.pop(), keys_and_values.pop()) {
-                (Some(value), Some(key)) => Some((key.to_string(), value.to_string())),
-                _ => None,
-            },
-        )
-        .filter(|(key, _)| !default_filter.contains(key))
-        // [("DB", "foo.db"), ("PORT", "99"), ("PATH", "/fake")]
-        .filter(|(key, _)| !filter_env_vars.contains(key))
-        // [("DB", "foo.db"), ("PORT", "99")]
-        .filter(|(key, _)| {
-            select_env_vars.is_empty()
-                || select_env_vars.contains("*")
-                || select_env_vars.contains(key)
-        })
-        // [("DB", "foo.db")]
-        .collect::<HashMap<_, _>>();
-
-    Ok(env_vars)
 }
 
 struct ClientConnectionHandler {
@@ -284,7 +227,7 @@ impl ClientConnectionHandler {
                 let environ_path = PathBuf::from("/proc").join(pid).join("environ");
 
                 let env_vars_result =
-                    select_env_vars(environ_path, env_vars_filter, env_vars_select).await;
+                    env::select_env_vars(environ_path, env_vars_filter, env_vars_select).await;
 
                 self.respond(DaemonMessage::GetEnvVarsResponse(env_vars_result))
                     .await?