+++ mirrord-layer/src/lib.rs
@@ -29,7 +29,7 @@ use libc::c_int;
 use mirrord_config::{
     config::MirrordConfig, pod::PodConfig, util::VecOrSingle, LayerConfig, LayerFileConfig,
 };
-use mirrord_macro::hook_guard_fn;
+use mirrord_macro::{hook_fn, hook_guard_fn};
 use mirrord_protocol::{
     dns::{DnsLookup, GetAddrInfoRequest},
     ClientCodec, ClientMessage, DaemonMessage, EnvVars, GetEnvVarsRequest,
@@ -91,12 +91,41 @@ pub(crate) static ENABLED_UDP_OUTGOING: OnceLock<bool> = OnceLock::new();
 
 pub(crate) const MIRRORD_SKIP_LOAD: &str = "MIRRORD_SKIP_LOAD";
 
+/// Check if we're running in NixOS or Devbox
+/// if so, add `sh` to the skip list because of https://github.com/metalbear-co/mirrord/issues/531
+fn nix_devbox_patch(config: &mut LayerConfig) {
+    let mut current_skip = config
+        .skip_processes
+        .clone()
+        .map(VecOrSingle::to_vec)
+        .unwrap_or_default();
+
+    if is_nix_or_devbox() && !current_skip.contains(&"sh".to_string()) {
+        current_skip.push("sh".into());
+        std::env::set_var("MIRRORD_SKIP_PROCESSES", current_skip.join(";"));
+        config.skip_processes = Some(VecOrSingle::Multiple(current_skip));
+    }
+}
+
+/// Check if NixOS or Devbox by discrimnating env vars.
+fn is_nix_or_devbox() -> bool {
+    if let Ok(res) = std::env::var("IN_NIX_SHELL") && res.as_str() == "1" {
+        true
+    }
+    else if let Ok(res) = std::env::var("DEVBOX_SHELL_ENABLED") && res.as_str() == "1" {
+        true
+    } else {
+        false
+    }
+}
+
 #[ctor]
 fn before_init() {
     if !cfg!(test) {
         if let Ok(Ok(true)) = std::env::var(MIRRORD_SKIP_LOAD).map(|value| value.parse::<bool>()) {
             return;
         }
+
         let args = std::env::args().collect::<Vec<_>>();
         let given_process = args.first().unwrap().split('/').last().unwrap();
 
@@ -108,7 +137,8 @@ fn before_init() {
             .generate_config();
 
         match config {
-            Ok(config) => {
+            Ok(mut config) => {
+                nix_devbox_patch(&mut config);
                 let skip_processes = config.skip_processes.clone().map(VecOrSingle::to_vec);
 
                 if should_load(given_process, skip_processes) {
@@ -472,6 +502,13 @@ fn enable_hooks(enabled_file_ops: bool, enabled_remote_dns: bool) {
 
     unsafe {
         let _ = replace!(&mut interceptor, "close", close_detour, FnClose, FN_CLOSE);
+        let _ = replace!(
+            &mut interceptor,
+            "close$NOCANCEL",
+            close_nocancel_detour,
+            FnClose_nocancel,
+            FN_CLOSE_NOCANCEL
+        );
     };
 
     unsafe { socket::hooks::enable_socket_hooks(&mut interceptor, enabled_remote_dns) };
@@ -519,6 +556,12 @@ pub(crate) unsafe extern "C" fn close_detour(fd: c_int) -> c_int {
     } else {
         FN_CLOSE(fd)
     }
+}
+
+// no need to guard because we call another detour which will do the guard for us.
+#[hook_fn]
+pub(crate) unsafe extern "C" fn close_nocancel_detour(fd: c_int) -> c_int {
+    close_detour(fd)
 }
 
 #[cfg(test)]